Auto-switching stats
Full implementation. Changed the select drops in the skills tab to be spans, and then implemented the JS conditional using some ternary operators. Could of course swap to if statements or case statements, but I felt that the ternary ended up being cleaner by just being fewer lines of code.